merge master

+++ ProjetBD2018.py
@@ -1,6 +1,5 @@
 import os
-from flask import Flask, render_template, request, redirect, url_for
-from werkzeug.exceptions import RequestEntityTooLarge
+from flask import Flask, render_template, request, redirect, send_from_directory
 from werkzeug.utils import secure_filename
 import hashlib
 import re
@@ -26,6 +25,11 @@ app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 app.config['MAX_CONTENT_LENGTH'] = 20 * 1024 * 1024  # 20 MB
 
 
+@app.route('/images/<filename>')
+def uploaded_file(filename):
+    return send_from_directory(app.config['UPLOAD_FOLDER'], filename)
+
+
 @app.route('/')
 def accueil():
     requete = 'SELECT id, nom From Biere'
@@ -47,6 +51,9 @@ def nous_contacter_post():
 
 @app.route('/ajout-biere', methods=['GET'])
 def ajout_biere():
+    global model
+    model['message_erreur'] = ""
+
     if 'est_une_microbrasserie' not in model['utilisateur_courant'].keys() or not model['utilisateur_courant']['est_une_microbrasserie']:
         return redirect('/')
 
@@ -62,35 +69,66 @@ def ajout_biere():
 
 @app.route('/ajout-biere', methods=['POST'])
 def ajout_biere_post():
+    model['message_erreur'] = ""
+    prix_ = request.form.get('prix')
+    alcool_ = request.form.get('alcool')
+    ibu_ = request.form.get('ibu')
+    annee_ = request.form.get('annee')
+    id_sorte_ = request.form.get('id_sous_sorte')
+    try:
+        if prix_ and alcool_ and ibu_ and annee_ and id_sorte_:
+            prix = float(prix_)
+            pourcentage_alcool = float(alcool_)
+            ibu = int(ibu_)
+            annee_production = int(annee_)
+            id_sorte = int(id_sorte_)
+        else:
+            raise ValueError
+    except ValueError:
+        model['message_erreur'] = "Veuillez entrer des nombres valides"
+        return render_template('ajout-biere.html', model=model)
 
-    prix = request.form.get('prix')
     nom = request.form.get('nom')
-    pourcentage_alcool = request.form.get('alcool')
-    ibu = request.form.get('ibu')
-    annee_production = request.form.get('annee')
     description = request.form.get('description')
-    id_sorte = request.form.get('id_sous_sorte')
-    image = request.form.get('image')
+    if not nom or len(nom) == 0:
+        model['message_erreur'] = "Vous devez saisir un nom"
+        return render_template('ajout-biere.html', model=model)
+    if not re.match(r"[1-9a-zA-Z\s\-]+$", nom):
+        model['message_erreur'] = "Vous devez saisir un nom valide"
+        return render_template('ajout-biere.html', model=model)
+    requete = 'SELECT nom FROM Biere WHERE nom=%s'
+    nom_existe = bd.execute_requete_lecture(requete, nom)
+    if nom_existe:
+        model['message_erreur'] = "Une bière portant ce nom existe déjà dans le système"
+        return render_template('ajout-biere.html', model=model)
+
+    if not description or len(description) == 0:
+        model['message_erreur'] = "Vous devez saisir une description"
+        return render_template('ajout-biere.html', model=model)
+    if not re.match(r"[1-9a-zA-Z\s\-]+$", description):
+        model['message_erreur'] = "Veuillez saisir une description valide"
+        return render_template('ajout-biere.html', model=model)
 
     if 'image' in request.files:
         file = request.files['image']
         if file and file.filename != '' and _allowed_file(file.filename):
             filename = secure_filename(file.filename)
-            path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
+            path = os.path.join(app.config['UPLOAD_FOLDER'], str(datetime.datetime.now()).replace(' ', '') + filename)
             file.save(path)
             file_length = os.stat(path).st_size
             if file_length > 5000000:  # 5 MB
                 os.remove(path)
-                model['message_erreur_upload'] = "La taille de l'image doit etre de 5 Mo maximum"
+                model['message_erreur'] = "La taille de l'image doit etre de 5 Mo maximum"
                 return render_template('ajout-biere.html', model=model)
-            else:
-                pass
+            else:  # Tout est valide
+                id_microbrasserie = model['utilisateur_courant']['id']
+                date_ajout = datetime.datetime.now().strftime('%Y-%m-%d')
+                requete = 'INSERT INTO Biere (image_url, prix, nom, pourcentage_alcool, ibu, annee_production, description, id_microbrasserie, date_ajout, id_sorte) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s);'
+                bd.execute_requete_ecriture(requete, (path, prix, nom, pourcentage_alcool, ibu, annee_production, description, id_microbrasserie, date_ajout, id_sorte))
+                model['message_reussite'] = "La bière a été importé avec succès"
+                return render_template('ajout-biere.html', model=model)
 
-    # 'image_url'
-    # 'id_microbrasserie'
-    # 'date_ajout'
-
-    model['message_erreur_upload'] = "L'image n'a pas pu etre importe"
+    model['message_erreur'] = "L'image n'a pas pu etre importe"
     return render_template('ajout-biere.html', model=model)
 
 
@@ -124,13 +162,13 @@ def bieres():
             sous_sortes = bd.execute_requete_lecture(requete, id_sorte, fetchall=True, obtenir_dict=True)
 
             # Bieres dispos
-            requete = 'SELECT B.id AS id, B.nom AS bnom, S.nom AS snom, B.prix AS prix, B.ibu AS ibu, B.pourcentage_alcool AS pa FROM Biere B, Sorte S WHERE B.id_sorte IN (SELECT id_sorte_enfant FROM Type_de WHERE id_sorte_parent = %s) AND B.id_sorte = S.id;'
+            requete = 'SELECT B.image_url, B.id AS id, B.nom AS bnom, S.nom AS snom, B.prix AS prix, B.ibu AS ibu, B.pourcentage_alcool AS pa FROM Biere B, Sorte S WHERE B.id_sorte IN (SELECT id_sorte_enfant FROM Type_de WHERE id_sorte_parent = %s) AND B.id_sorte = S.id;'
             biere_dispo = bd.execute_requete_lecture(requete, id_sorte, fetchall=True, obtenir_dict=True)
 
         except (ValueError, TypeError):
             id_sorte = None
             sous_sortes = []
-            requete = 'SELECT B.id AS id, B.nom AS bnom, S.nom AS snom, B.prix AS prix, B.ibu AS ibu, B.pourcentage_alcool AS pa FROM Biere B, Sorte S WHERE B.id_sorte = S.id;'
+            requete = 'SELECT B.image_url, B.id AS id, B.nom AS bnom, S.nom AS snom, B.prix AS prix, B.ibu AS ibu, B.pourcentage_alcool AS pa FROM Biere B, Sorte S WHERE B.id_sorte = S.id;'
             biere_dispo = bd.execute_requete_lecture(requete, fetchall=True, obtenir_dict=True)
 
     _handle_erreur_connexion()
@@ -160,7 +198,7 @@ def connexion():
         if row is not None:
             id_, hash_bd = row
             if hash_bd is not None and hash_bd == hashlib.sha512(mot_de_passe.encode('utf-8')).digest():
-                requete = 'SELECT nom, prenom FROM Utilisateur WHERE id=%s;'
+                requete = 'SELECT nom, prenom, id FROM Utilisateur WHERE id=%s;'
                 model['utilisateur_courant'] = bd.execute_requete_lecture(requete, id_, obtenir_dict=True)
                 model['utilisateur_courant']['panier'] = dict()
                 model['utilisateur_courant']['nombre_bieres'] = 0
@@ -196,7 +234,7 @@ def signup():
         if mot_de_passe == confirmation and re.match(r"[^@\s]+@[^@\s]+\.[a-zA-Z0-9]+$", courriel) and re.match(r'[a-zA-Z\s\-]+$', prenom) and re.match(r'[a-zA-Z\s\-]+$', nom):
             hash_bd = hashlib.sha512(mot_de_passe.encode('utf-8')).digest()
             ajout = 'INSERT INTO Utilisateur (courriel, nom, prenom) VALUES (%s, %s, %s);'
-            bd.execute_requete_ecriture(ajout, courriel, nom, prenom)
+            bd.execute_requete_ecriture(ajout, (courriel, nom, prenom))
             ajout_mdp = 'INSERT INTO Mot_de_passe (mot_de_passe) VALUES (%s);'
             bd.execute_requete_ecriture(ajout_mdp, hash_bd)
             return redirect('/')
